pyproject.toml
[project]
name = "pg-context-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "mysql-connector-python>=9.4.0",
    "orjson>=3.11.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "temporalio>=1.16.0",
    "uvicorn>=0.35.0",
]
